[app/registers] extracting target register tests
